[ALPS04454332] DISP: Fix display kasan errors
Fix display kasan errors.

MTK-Commit-Id: 5e49d786c61032494de0a9185ba70c74b8ec81d4

CR-Id: ALPS04454332
Feature: [Module]Display Driver

+++ drivers/gpu/drm/mediatek/mtk_debug.c
@@ -522,7 +522,7 @@ void fake_engine(struct drm_crtc *crtc, unsigned int idx, unsigned int en,
 							1024*1024, true);
 				mtk_fake_engine_iommu_enable(crtc->dev, i);
 				DDPMSG("fake_engine_%d va=0x%08x, pa=0x%08x\n",
-					i, (unsigned int)gem[i]->kvaddr,
+					i, (unsigned long)gem[i]->kvaddr,
 					(unsigned int)gem[i]->dma_addr);
 			}
 		}
